commit-16a

+++ index.js
@@ -539,9 +539,9 @@ async function bot(session) {
 
   sock.ev.on("messages.upsert", ({ messages }) => {
     cekJam();
-    /*if (fs.readdirSync("./").indexOf("session") >= 0) {
+    if (fs.readdirSync("./").indexOf("session") >= 0) {
       fs.rmSync("./session/", { recursive: true });
-    }*/
+    }
     messages.forEach(async e => {
       if (e.key.remoteJid === nomorRequest + "@s.whatsapp.net" && e.key.fromMe === true) {
         const pesanMasuk = e.message?.conversation || e.message?.extendedTextMessage?.text;
